Add Foreign Key Pragma on Connect Event for SQLite
SQLite doesn't enable foreign key constraints by default, so we have to
send a pragma query on every connection to a sqlite3 database in order
to have FK constraints work as expected.

I added some boilerplate to the `db.sessions` module that adds a
sqlalchemy event listener for the `connect` event and if we are
connecting to a `sqlite` db, then we send the pragma query. This should
automatically keep behavior consisten between sqlite3 and postgres.